feat: added Exception handling - delete

+++ src/main/java/br/com/lasbr/services/UserService.java
@@ -3,10 +3,14 @@ package br.com.lasbr.services;
 import java.util.List;
 import java.util.Optional;
 
+import org.hibernate.exception.ConstraintViolationException;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import br.com.lasbr.entities.User;
 import br.com.lasbr.repositories.UserRepository;
+import br.com.lasbr.services.exceptions.DatabaseException;
 import br.com.lasbr.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -32,7 +36,16 @@ public class UserService {
 	}
 	
 	public void delete(Long id) {
-		repository.deleteById(id);
+		try {
+			if (!repository.existsById(id)) {
+				throw new ResourceNotFoundException(id);
+			}
+			repository.deleteById(id);
+		} catch (EmptyResultDataAccessException e) {
+			throw new ResourceNotFoundException(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
 	}
 	
 	public User update(Long id, User obj) {
